refactor arrayAddOns sorting functions into helper
